pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling" ]

[project]
name = "precise-patterns"
version = "0.0.1-alpha"
description = "The next iteration of Stock Patterns"
readme = "README.md"
keywords = [ "algo-trading", "patterns", "stock-market" ]
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: Financial and Insurance Industry",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Office/Business :: Financial :: Investment",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [ "pyee==13", "python-dateutil==2.9.0.post0" ]
optional-dependencies.docs = [
  "furo==2025.9.25",
  "sphinx==8.2.3",
]
author = "Benny Thadikaran"
